版本 1.3.0

See merge request xugu-open/dbeaver!2

+++ plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/exec/DBCStatement.java
@@ -1,6 +1,6 @@
 /*
  * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2020 DBeaver Corp and others
+ * Copyright (C) 2010-2021 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -97,7 +97,7 @@ public interface DBCStatement extends DBPObject, DBRBlockingObject, DBPCloseable
      * @return number of row updated
      * @throws DBCException on error
      */
-    int getUpdateRowCount() throws DBCException;
+    long getUpdateRowCount() throws DBCException;
 
     /**
      * Checks whether there are additional results (result set or update count).
